Implement non-interactive mode
This allows scripts to be read from a file, a non-terminal stdin, or a
command-line argument (via the `-c` flag). The new flag allows Mesh to
be used as a user's login shell on Ubuntu - GDM launches `gnome-session`
with `$SHELL -c '/usr/bin/gnome-session -l --systemd --session=ubuntu'`.
Without the flag, it will fail to launch, and immediately return to the
GDM login screen.