Add support for multi-line strings
Also, run the lexer in the main goroutine, and run the parser on a
separate goroutine. This is the opposite of how it was previously. The
scanner and lexer run in lock-step, so it makes sense to run them
together. And in the case of multi-line inputs, the parser might still
be in the middle of a recursive descent when we need to get a new line,
so it's easier if this is running concurrently.